Upgrade to research-grade infrastructure (v0.1.0)
Restructure engine into clean modules (policy, metrics, io, accounts)
separated from UI. Add YAML config system, CSV/Parquet-ready data export,
run artifact saving with metadata (seed, git hash, params), and per-phase
profiling hooks. Add comprehensive test suite: strict accounting invariants,
property-based tests (hypothesis), smoke tests, and CI workflow. Add ruff
linting, mypy type checking, and GitHub Actions. Fix pre-existing
test_employment_dynamics failure.

New modules: engine/policy.py, engine/metrics.py, engine/io.py,
engine/accounts.py. New docs: MODEL_ASSUMPTIONS.md, CONTRIBUTING.md,
CHANGELOG.md. New config: configs/baseline.yaml, pyproject.toml.

+++ engine/production.py
@@ -7,9 +7,6 @@ production function:  output = productivity * labor^alpha * capital^beta
 
 from __future__ import annotations
 
-import math
-from typing import Dict
-
 
 def cobb_douglas(
     productivity: float,
@@ -25,7 +22,7 @@ def cobb_douglas(
     """
     if labor <= 0 or capital <= 0:
         return 0.0
-    return productivity * (labor ** labor_share) * (capital ** capital_share)
+    return float(productivity * (labor ** labor_share) * (capital ** capital_share))
 
 
 def desired_labor(
@@ -42,7 +39,7 @@ def desired_labor(
     if capital_factor <= 0:
         return 0.0
     raw = target_output / capital_factor
-    return raw ** (1.0 / labor_share)
+    return float(raw ** (1.0 / labor_share))
 
 
 def firm_target_output(
